Timeout added

+++ messageSender/wasender.py
@@ -10,9 +10,13 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 
 class WASender(Sender):
-    def wait_for_element(self, selector, count):
-        while len(self.driver.execute_script(f'return document.querySelectorAll(\'{selector}\')')) < count:
+    def wait_for_element(self, selector, count) -> bool:
+        save_wait = 60 * 10
+        while len(self.driver.execute_script(f'return document.querySelectorAll(\'{selector}\')')) < count and save_wait > 0:
             time.sleep(0.1)
+            save_wait = save_wait - 1
+        print(save_wait > 0)
+        return save_wait > 0
 
     def __init__(self, options: webdriver.ChromeOptions = webdriver.ChromeOptions(), driver_class = webdriver.Chrome, profile_path = os.path.join(os.getcwd(), "profile")):
         super().__init__()
@@ -28,7 +32,10 @@ class WASender(Sender):
 
         self.driver = driver_class(options=options)
         self.driver.get(constants.WA_URL)
-        self.wait_for_element("#side", 1)
+
+        if not self.wait_for_element("#side", 1):
+            raise ValueError
+
         self.driver.find_element(By.CSS_SELECTOR, "body").send_keys(Keys.ESCAPE)
 
         self.current = None
@@ -41,9 +48,13 @@ class WASender(Sender):
     def send(self) -> None:
         self.driver.execute_script("document.querySelectorAll(\"footer button\")[2].click();")
 
-    def send_text(self, to: str, text: str):
+    def send_text(self, to: str, text: str) -> bool:
         self.__go_to_user(to)
-        self.wait_for_element("footer button", 3)
+
+        if not self.wait_for_element("footer button", 3):
+            logger.collect_log(f"text is not sended for {to}")
+            return False
+
         self.driver.execute_script(
             f"const text = `{text}`;"
             "const dataTransfer = new DataTransfer();"
@@ -56,6 +67,7 @@ class WASender(Sender):
         )
         self.send()
         self.waiter()
+        return True
 
     def quit(self):
         time.sleep(1)
@@ -63,38 +75,45 @@ class WASender(Sender):
     
     def send_image(self, to, image_path):
         self.__go_to_user(to)
-        self.wait_for_element('[data-icon=\"plus-rounded\"]', 1)
+
+        if not self.wait_for_element('[data-icon=\"plus-rounded\"]', 1):
+            logger.collect_log(f"image {image_path} not sended for {to}")
+            return False
         
         self.driver.execute_script("document.querySelector(\"[data-icon='plus-rounded']\").click()")
-        count = self.driver.execute_script(
-            'return document.querySelectorAll("span + div > span[data-icon=\'msg-dblcheck\']").length;'
-        )
+
         self.driver.find_element(
             By.CSS_SELECTOR,
             'input[accept="image/*,video/mp4,video/3gpp,video/quicktime"]'
         ).send_keys(
             image_path
         )
-        self.wait_for_element("div:has(+input) div[role=\"button\"]:has(> span > svg)", 2)
+
+        if not self.wait_for_element("div:has(+input) div[role=\"button\"]:has(> span > svg)", 2):
+            logger.collect_log(f"image {image_path} not sended for {to}")
+            return False
+
         self.driver.execute_script(
             'document.querySelectorAll("div:has(+input) div[role=\'button\']:has(> span > svg)")[1].click()'
         )
-        self.wait_for_element('span + div > span[data-icon="msg-dblcheck"]', count)
+
         self.waiter()
         os.remove(image_path)
+        return True
 
-    @staticmethod
-    def __await_func(wait_for: str) -> str:
-        return " async function waiter(wait_for) {" \
+    def __await_func(self, waits_for: list[str]) -> bool:
+        function_text =  " async function waiter(waits_for) {" \
         "     while (true) {" \
         "        let l = document.querySelectorAll(\"div[role='application'] div[role='row']\");" \
-        "         let el = l[l.length - 1];" \
+        "        let el = l[l.length - 1];" \
         "        if (el.querySelector(\"span+div>span[aria-hidden='false']\") === null) {await delay(); continue}" \
         "        let attr = el.querySelector(\"span+div>span[aria-hidden='false']\").getAttribute(\"data-icon\");" \
-        "         if (wait_for === attr){" \
-        "             return" \
-        "         }" \
-        "         await delay();" \
+        "        for (let index = 0; index < waits_for.length; index++) {" \
+        "           if (waits_for[index] === attr){" \
+        "              return" \
+        "           }" \
+        "        }" \
+        "        await delay();" \
         "     }" \
         " };" \
         " function delay() {" \
@@ -104,13 +123,14 @@ class WASender(Sender):
         "         }, 10);" \
         "     });" \
         " };" \
-        f"await waiter(\"{wait_for}\");"
+        f"await waiter(['{'\', \''.join(waits_for)}']);"
+        try:
+            self.driver.execute_script(function_text)
+        except Exception as e:
+            logger.collect_log(str(e))
+            return False
+        return True
     
     def waiter(self):
-        try:
-            self.driver.execute_script(self.__await_func("msg-time"))
-        except Exception as e:
-            logger.collect_log(e)
-
-        self.driver.execute_script(self.__await_func("msg-dblcheck"))
-        time.sleep(0.1)
+        self.__await_func(["msg-time"])
+        self.__await_func(["msg-check", "msg-dblcheck"])
